+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.0.21
+
+- Remove navigate service
+- Remove exit navigation service
+- Add Remote entity allowing to control and turn off control using a virtual joystick
+
 ## v1.0.20
 
 - Fix navigation service schema [#223](https://github.com/sh00t2kill/dolphin-robot/issues/223)